backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
supabase==2.11.0
pydantic==2.10.4
openai==1.59.3
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.159.0
twilio==9.4.0
resend==2.11.0
deepgram-sdk==4.8.0
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1